Feature/get base route id

+++ app/src/main/java/com/timi/seulseul/data/model/request/AlarmPostRequest.kt
@@ -1,7 +1,7 @@
 package com.timi.seulseul.data.model.request
 
 data class AlarmPostRequest(
-    val base_route_id : Int,
+    val base_route_id : Long,
     val alarmTime : Int,
     val alarmTerm : Int
 )
